sequelx_demo/backend/requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
pydantic>=2.9.0
